tools/weather_tool.py
import requests
import urllib3

# Suppress SSL warnings
urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)

# City to latitude & longitude mapping
CITY_COORDS = {
    "Delhi": (28.6139, 77.2090),
    "Mumbai": (19.0760, 72.8777),
    "Goa": (15.2993, 74.1240),
    "Bangalore": (12.9716, 77.5946),
    "Chennai": (13.0827, 80.2707),
    "Hyderabad": (17.3850, 78.4867),
    "Kolkata": (22.5726, 88.3639),
    "Jaipur": (26.9124, 75.7873)
}

def get_weather(city):
    """
    Get daily weather forecast for a city using Open-Meteo API.
    """
    # Make city search case-insensitive
    city_title = city.title()
    
    if city_title not in CITY_COORDS:
        return {"message": "City not supported for weather"}

    lat, lon = CITY_COORDS[city_title]

    url = (
        "https://api.open-meteo.com/v1/forecast"
        f"?latitude={lat}&longitude={lon}"
        "&daily=temperature_2m_max,temperature_2m_min"
        "&timezone=auto"
    )

    try:
        response = requests.get(url, timeout=10, verify=False)
        response.raise_for_status()
        data = response.json()

        forecast = []
        dates = data["daily"]["time"]
        max_temps = data["daily"]["temperature_2m_max"]
        min_temps = data["daily"]["temperature_2m_min"]

        for i in range(len(dates)):
            forecast.append({
                "date": dates[i],
                "min_temp": min_temps[i],
                "max_temp": max_temps[i]
            })

        return forecast
    except Exception as e:
        # Return dummy data if API fails
        import datetime
        today = datetime.date.today()
        return [
            {
                "date": str(today + datetime.timedelta(days=i)),
                "min_temp": 20 + i,
                "max_temp": 30 + i
            }
            for i in range(7)
        ]


# Test the tool
if __name__ == "__main__":
    result = get_weather("Delhi")
    for day in result[:3]:
        print(day)
