requirements.txt
streamlit>=1.28.0
requests>=2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
tensorflow>=2.13.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
